Update OHW interface to version 3. Use common ABI description file with Qemu for both Sparc32 and Sparc64. Remove private definitions and magic constants.
git-svn-id: svn://coreboot.org/openbios/openbios-devel@176 f158a5a8-5612-0410-a976-696ce0be7e32

+++ arch/sparc64/boot.c
@@ -16,10 +16,10 @@ int linux_load(struct sys_info *, const char *filename, const char *cmdline);
 void boot(void);
 
 struct sys_info sys_info;                                                       
-uint32_t kernel_image;
-uint32_t kernel_size;
-uint32_t cmdline;
-uint32_t cmdline_size;
+uint64_t kernel_image;
+uint64_t kernel_size;
+uint64_t cmdline;
+uint64_t cmdline_size;
 char boot_device;
 
 void boot(void)
